Issue #3738: extract AbstractScalaCodegen.
This seems to not change anything in the output, at least for the Petstore samples.
